Moved some methods to test utils.
